Add module:stream manipulation remote commands to Web UI.
Add pages to the single client and SSM page trees to allow adding,
changing, and removing modules and streams.

SSM filters by base channel tree, and splits and orders execution
steps to account for the fact that some operations steps may not apply
to all clients.

Executed on the client via remote commands, so the script.run
privilege is required.

A new view has been added to the schema to support the above.

OLDIS-6915

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>