Fix spacewalk-report actions le/ge where clause.
ORA-01843 was being thrown because the timestamp displayed to the user
was not formatted correctly for use in a where clause. The customer
would typically copy/paste the timestamp.

Correct the formatting of the timestamp displayed to the user to be
compatible with a where clause, as per the other reports.

Orabug: 34646950

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>