Fix HTTP 500 and ORA-01830 on client scap report
Timestamps submitted by client SCAP reports can be in Timestamp format
instead of date, and have milliseconds appended. This doesn't match
the timestamp fstring used in the SCAP backend SQL.

This causes ORA-01830 to be logged in xmlrpc log and HTTP 500 to be
returned to the client.

Rework sanitisation to remove the milliseconds.

Orabug: 34823889

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>