Clone comps.xml when cloning channels.
When cloning channels the comps.xml file was not cloned, instead a
pointer to the source channel's file was created in the database. The
clone would function correctly until the source channel was synced and
included an update to the groups detailed in comps.xml. At this point
the comps.xml and clone channel package/errata list would be out of
sync, resulting in yum reporting "Downloading successful, but checksum
doesn't match".

This change copies the comps.xml when cloning channels and makes the
database record point at the copied file. This is the same as was done
for modules.yaml.

Both modules.yaml and comps.xml files are now copied when cloning a
channel.

reposync.py and spacewalk-backend.spec now set comps dir
group/permissions on first sync or for existing files on
spacewalk-backend RPM upgrade.

Orabug: 34819789

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>